+++ src/greptimeai/patchee/openai_patchee/chat_completion.py
