+++ src/greptimeai/patchee/openai_patchee/completion.py
