+++ src/greptimeai/patchee/openai_patchee/embedding.py
