+++ src/greptimeai/patchee/openai_patchee/image.py
