+++ src/greptimeai/patchee/openai_patchee/model.py
